Se validan los txts para que acepten valores double

+++ RandomVarGenerator/RandomVarGenerator/UniformGenerator.cs
@@ -9,8 +9,9 @@ namespace RandomVarGenerator
 {
     class UniformGenerator
     {
-        public int a { get; set; }
-        public int b { get; set; }
+        public double a { get; set; }
+        public double b { get; set; }
+        
         public Generator congruentialGenerator = new Generator() { seed = 31767, a = 71561, c = 56822, M = 341157 };
 
         public double Generate()
